+++ tests/__tests__/components/Stubs.vue
