+++ tests/__tests__/stubs.js
@@ -1,10 +1,10 @@
 import { render, cleanup } from '@testing-library/vue'
-import Form from './components/Form'
+import Stubs from './components/Stubs'
 
 afterEach(cleanup)
 
 test('Form contains search button', () => {
-  const { getByText } = render(Form, {
+  const { getByText } = render(Stubs, {
     stubs: ['FontAwesomeIcon']
   })
   getByText('Search now')
